Buffered request first introduction pcserver

+++ pcserver/main.h
@@ -3,7 +3,7 @@
 
 void printBufferHex(const char* buffer, size_t length);
 
-void sampleChannels(int serialfd, bool* sampleChannels, int numChannels, unsigned int samplingFrequency, unsigned int time, uint8_t operatingMode);
+void sampleChannels(int serialfd, bool* sampleChannels, int numChannels, unsigned int samplingFrequency, unsigned int time, unsigned int operatingMode);
 
 void continuosSampling(int serialfd, bool* sampleChannels, int numChannels, unsigned int samplingFrequency, unsigned int time);
